refactor: merge command/idle handlers back into tui_context.rs
Keep tui_handlers_inference.rs separate (borrow-checker free fns are
structurally different), but merge commands + idle back into tui_context.rs
for cohesion — they're all methods on the same struct.

Result: 2 files instead of 4
- tui_context.rs (1,249 lines) — struct + init + event loop + commands + idle
- tui_handlers_inference.rs (480 lines) — inference loop + borrow-split fns

Over 600 lines but cohesive. The small methods keep it navigable.
