+++ koda-cli/src/tui_context.rs
@@ -6,28 +6,31 @@
 
 use crate::input;
 use crate::sink::UiEvent;
+use crate::tui_commands::{self, SlashAction};
+use crate::tui_output;
 use crate::tui_render::TuiRenderer;
 use crate::tui_types::{
     MIN_VIEWPORT_HEIGHT, MenuContent, PromptMode, ProviderWizard, Term, TuiState,
 };
 use crate::tui_viewport::{
-    draw_viewport, emit_above, init_terminal, maybe_resize_viewport,
-    restore_terminal,
+    draw_viewport, drain_pending_resizes, emit_above, init_terminal,
+    maybe_resize_viewport, restore_terminal, scroll_past_and_reinit,
 };
 
 use anyhow::Result;
-use crossterm::event::EventStream;
+use crossterm::event::{Event, EventStream, KeyCode, KeyModifiers};
 use futures_util::StreamExt;
 use koda_core::agent::KodaAgent;
 use koda_core::approval::{self, ApprovalMode};
 use koda_core::config::KodaConfig;
+use koda_core::db::Role;
 use koda_core::engine::EngineCommand;
 use koda_core::persistence::Persistence;
 use koda_core::providers::LlmProvider;
 use koda_core::session::KodaSession;
 use ratatui::{
     style::{Color, Modifier, Style},
-    text::Line,
+    text::{Line, Span},
 };
 use ratatui_textarea::TextArea;
 use std::collections::VecDeque;
@@ -83,6 +86,20 @@ pub(crate) struct TuiContext {
     pub shared_mode: approval::SharedMode,
     pub agent: Arc<KodaAgent>,
     pub project_root: PathBuf,
+}
+
+/// Outcome of dispatching a single command.
+pub(crate) enum CommandOutcome {
+    /// No input was available to dispatch.
+    NoInput,
+    /// Command was fully handled (slash command, dropdown opened, etc.).
+    Handled,
+    /// Non-slash input — ready to start an inference turn.
+    StartInference {
+        pending_images: Option<Vec<koda_core::providers::ImageData>>,
+    },
+    /// /quit was invoked.
+    Quit,
 }
 
 impl TuiContext {
@@ -298,16 +315,7 @@ impl TuiContext {
         crate::startup::print_resume_hint(&self.session.id);
     }
 
-    /// The main event loop. Dispatches queued commands, handles inference
-    /// turns, and processes idle keyboard input.
-    ///
-    /// Channels stay in the caller (`run()`) and are passed in because
-    /// they're consumed differently by `tokio::select!`.
-    ///
-    /// The heavy lifting is delegated to handler modules:
-    /// - `tui_handlers_commands`  — slash commands and inference preparation
-    /// - `tui_handlers_inference` — inference turn inner loop + cleanup
-    /// - `tui_handlers_idle`      — idle keyboard/paste/menu handling
+    /// The main event loop.
     pub async fn run_event_loop(
         &mut self,
         ui_tx: &mpsc::Sender<UiEvent>,
@@ -315,21 +323,17 @@ impl TuiContext {
         cmd_tx: &mpsc::Sender<EngineCommand>,
         cmd_rx: &mut mpsc::Receiver<EngineCommand>,
     ) -> Result<()> {
-        use crate::tui_handlers_commands::CommandOutcome;
-
         loop {
             if self.should_quit {
                 break;
             }
 
-            // ── Dispatch queued / pending commands ───────────────
+            // ── Dispatch queued / pending commands ───────────
             if self.tui_state == TuiState::Idle {
                 if let Some(raw) = self.dequeue_input() {
                     match self.dispatch_command(&raw).await {
                         CommandOutcome::NoInput => {}
-                        CommandOutcome::Handled => {
-                            continue;
-                        }
+                        CommandOutcome::Handled => continue,
                         CommandOutcome::Quit => {
                             self.should_quit = true;
                             continue;
@@ -339,25 +343,870 @@ impl TuiContext {
                                 pending_images, ui_tx, ui_rx, cmd_tx, cmd_rx,
                             )
                             .await?;
-                            // Loop back to drain queue before blocking on keyboard
                             continue;
                         }
                     }
                 }
             }
 
-            // Redraw viewport
             self.draw()?;
 
-            // ── Idle: wait for keyboard input ────────────────────
+            // ── Idle: wait for keyboard input ────────────────
             tokio::select! {
                 Some(Ok(ev)) = self.crossterm_events.next() => {
                     self.handle_idle_event(ev).await?;
                 }
             }
         }
-
         Ok(())
+    }
+
+    // ═══════════════════════════════════════════════════════════
+    // Command dispatch (slash commands, dropdown openers, inference prep)
+    // ═══════════════════════════════════════════════════════════
+
+    /// Dequeue the next pending or queued input string, if any.
+    fn dequeue_input(&mut self) -> Option<String> {
+        if let Some(cmd) = self.pending_command.take() {
+            return Some(cmd);
+        }
+        if let Some(queued) = self.input_queue.pop_front() {
+            let mode = approval::read_mode(&self.shared_mode);
+            let icon = match mode {
+                ApprovalMode::Confirm => "🔒",
+                ApprovalMode::Auto => "⚡",
+            };
+            emit_above(
+                &mut self.terminal,
+                Line::from(vec![
+                    Span::styled(format!("{icon}> "), Style::default().fg(Color::Cyan)),
+                    Span::raw(queued.clone()),
+                ]),
+            );
+            return Some(queued);
+        }
+        None
+    }
+
+    /// Dispatch a raw input string.
+    async fn dispatch_command(&mut self, raw: &str) -> CommandOutcome {
+        let input = raw.trim().to_string();
+        if input.is_empty() {
+            return CommandOutcome::NoInput;
+        }
+        if input.starts_with('/') {
+            return self.dispatch_slash(&input).await;
+        }
+        self.prepare_inference_start(&input).await
+    }
+
+    async fn dispatch_slash(&mut self, input: &str) -> CommandOutcome {
+        if input.trim() == "/model" {
+            self.open_model_picker().await;
+            return CommandOutcome::Handled;
+        }
+        if input.trim() == "/provider" {
+            self.open_provider_picker();
+            return CommandOutcome::Handled;
+        }
+        if input.trim().starts_with("/provider ") {
+            let name = input.trim().strip_prefix("/provider ").unwrap().trim();
+            self.start_provider_wizard(name);
+            return CommandOutcome::Handled;
+        }
+        if input.trim() == "/sessions" {
+            self.open_session_picker().await;
+            return CommandOutcome::Handled;
+        }
+
+        let action = tui_commands::handle_slash_command(
+            &mut self.terminal,
+            input,
+            &mut self.config,
+            &self.provider,
+            &mut self.session,
+            &self.shared_mode,
+            &mut self.renderer,
+            &self.project_root,
+            &self.agent,
+            &mut self.pending_command,
+        )
+        .await;
+
+        match action {
+            SlashAction::Continue => {
+                self.reinit_after_slash().await;
+                CommandOutcome::Handled
+            }
+            SlashAction::Quit => {
+                tui_output::emit_line(
+                    &mut self.terminal,
+                    Line::styled("\u{1f43b} Goodbye!", Style::default().fg(Color::Cyan)),
+                );
+                CommandOutcome::Quit
+            }
+        }
+    }
+
+    // ── Dropdown openers ─────────────────────────────────────────
+
+    async fn open_model_picker(&mut self) {
+        let prov = self.provider.read().await;
+        match prov.list_models().await {
+            Ok(models) if !models.is_empty() => {
+                let items: Vec<crate::widgets::model_menu::ModelItem> = models
+                    .iter()
+                    .map(|m| crate::widgets::model_menu::ModelItem {
+                        id: m.id.clone(),
+                        is_current: m.id == self.config.model,
+                    })
+                    .collect();
+                let mut dd = crate::widgets::dropdown::DropdownState::new(
+                    items,
+                    "\u{1f43b} Select a model",
+                );
+                if let Some(idx) = dd.filtered.iter().position(|m| m.is_current) {
+                    dd.selected = idx;
+                    let max_vis = crate::widgets::dropdown::MAX_VISIBLE;
+                    if idx >= max_vis {
+                        dd.scroll_offset = idx + 1 - max_vis;
+                    }
+                }
+                self.menu = MenuContent::Model(dd);
+            }
+            Ok(_) => {
+                emit_above(
+                    &mut self.terminal,
+                    Line::styled(
+                        "  \u{26a0} No models available",
+                        Style::default().fg(Color::Yellow),
+                    ),
+                );
+            }
+            Err(e) => {
+                emit_above(
+                    &mut self.terminal,
+                    Line::styled(
+                        format!("  \u{2717} Failed to list models: {e}"),
+                        Style::default().fg(Color::Red),
+                    ),
+                );
+            }
+        }
+    }
+
+    fn open_provider_picker(&mut self) {
+        let providers = crate::repl::PROVIDERS;
+        let items: Vec<crate::widgets::provider_menu::ProviderItem> = providers
+            .iter()
+            .map(|(key, name, desc)| crate::widgets::provider_menu::ProviderItem {
+                key,
+                name,
+                description: desc,
+                is_current: koda_core::config::ProviderType::from_url_or_name("", Some(key))
+                    == self.config.provider_type,
+            })
+            .collect();
+        let mut dd = crate::widgets::dropdown::DropdownState::new(
+            items,
+            "\u{1f43b} Select a provider",
+        );
+        if let Some(idx) = dd.filtered.iter().position(|p| p.is_current) {
+            dd.selected = idx;
+            let max_vis = crate::widgets::dropdown::MAX_VISIBLE;
+            if idx >= max_vis {
+                dd.scroll_offset = idx + 1 - max_vis;
+            }
+        }
+        self.menu = MenuContent::Provider(dd);
+    }
+
+    fn start_provider_wizard(&mut self, name: &str) {
+        let ptype = koda_core::config::ProviderType::from_url_or_name("", Some(name));
+        let base_url = ptype.default_base_url().to_string();
+        let provider_name = ptype.to_string();
+
+        if ptype.requires_api_key() {
+            let env_name = ptype.env_key_name().to_string();
+            let has_key = koda_core::runtime_env::is_set(&env_name);
+            let label = if has_key {
+                format!("API key for {} (Enter to keep current)", ptype)
+            } else {
+                format!("API key for {}", ptype)
+            };
+            self.menu = MenuContent::WizardTrail(vec![("Provider".into(), provider_name)]);
+            self.prompt_mode = PromptMode::WizardInput {
+                label,
+                masked: true,
+            };
+            self.provider_wizard = Some(ProviderWizard::NeedApiKey {
+                provider_type: ptype,
+                base_url,
+                env_name,
+            });
+            self.textarea.select_all();
+            self.textarea.cut();
+        } else {
+            self.menu = MenuContent::WizardTrail(vec![("Provider".into(), provider_name)]);
+            self.prompt_mode = PromptMode::WizardInput {
+                label: format!("{} URL", ptype),
+                masked: false,
+            };
+            self.provider_wizard = Some(ProviderWizard::NeedUrl {
+                provider_type: ptype,
+            });
+            self.textarea.select_all();
+            self.textarea.cut();
+            self.textarea.insert_str(&base_url);
+        }
+    }
+
+    async fn open_session_picker(&mut self) {
+        match self.session.db.list_sessions(10, &self.project_root).await {
+            Ok(sessions) if !sessions.is_empty() => {
+                let items: Vec<crate::widgets::session_menu::SessionItem> = sessions
+                    .iter()
+                    .map(|s| crate::widgets::session_menu::SessionItem {
+                        id: s.id.clone(),
+                        short_id: s.id[..8.min(s.id.len())].to_string(),
+                        created_at: s.created_at.clone(),
+                        message_count: s.message_count,
+                        total_tokens: s.total_tokens,
+                        is_current: s.id == self.session.id,
+                    })
+                    .collect();
+                let mut dd = crate::widgets::dropdown::DropdownState::new(
+                    items,
+                    "\u{1f43b} Sessions",
+                );
+                if let Some(idx) = dd.filtered.iter().position(|s| s.is_current) {
+                    dd.selected = idx;
+                    let max_vis = crate::widgets::dropdown::MAX_VISIBLE;
+                    if idx >= max_vis {
+                        dd.scroll_offset = idx + 1 - max_vis;
+                    }
+                }
+                self.menu = MenuContent::Session(dd);
+            }
+            Ok(_) => {
+                emit_above(
+                    &mut self.terminal,
+                    Line::styled(
+                        "  No other sessions found.",
+                        Style::default().fg(Color::DarkGray),
+                    ),
+                );
+            }
+            Err(e) => {
+                emit_above(
+                    &mut self.terminal,
+                    Line::styled(
+                        format!("  \u{2717} Error: {e}"),
+                        Style::default().fg(Color::Red),
+                    ),
+                );
+            }
+        }
+    }
+
+    async fn reinit_after_slash(&mut self) {
+        self.viewport_height = MIN_VIEWPORT_HEIGHT;
+        self.crossterm_events = EventStream::new();
+        if let Ok(term) = init_terminal(self.viewport_height) {
+            self.terminal = term;
+        }
+        {
+            let prov = self.provider.read().await;
+            if let Ok(models) = prov.list_models().await {
+                self.completer
+                    .set_model_names(models.iter().map(|m| m.id.clone()).collect());
+            }
+        }
+        self.renderer.model = self.config.model.clone();
+        let _ = self.draw();
+    }
+
+    async fn prepare_inference_start(&mut self, input: &str) -> CommandOutcome {
+        let mut processed = input::process_input(input, &self.project_root);
+        processed.paste_blocks = std::mem::take(&mut self.paste_blocks);
+
+        for (i, _img) in processed.images.iter().enumerate() {
+            emit_above(
+                &mut self.terminal,
+                Line::from(vec![
+                    Span::raw("  "),
+                    Span::styled(
+                        format!("\u{1f5bc} Image {}", i + 1),
+                        Style::default().fg(Color::Magenta),
+                    ),
+                ]),
+            );
+        }
+
+        let mut user_message =
+            if let Some(context) = input::format_context_files(&processed.context_files) {
+                for f in &processed.context_files {
+                    emit_above(
+                        &mut self.terminal,
+                        Line::from(vec![
+                            Span::raw("  "),
+                            Span::styled(
+                                format!("\u{1f4ce} {}", f.path),
+                                Style::default().fg(Color::Cyan),
+                            ),
+                        ]),
+                    );
+                }
+                format!("{}\n\n{context}", processed.prompt)
+            } else {
+                processed.prompt.clone()
+            };
+
+        if let Some(pasted) = input::format_paste_blocks(&processed.paste_blocks) {
+            user_message = format!("{user_message}\n\n{pasted}");
+        }
+
+        if let Err(e) = self
+            .session
+            .db
+            .insert_message(&self.session.id, &Role::User, Some(&user_message), None, None, None)
+            .await
+        {
+            tracing::warn!("Failed to persist user message: {e}");
+        }
+
+        let pending_images = if processed.images.is_empty() {
+            None
+        } else {
+            Some(processed.images)
+        };
+
+        self.session.mode = approval::read_mode(&self.shared_mode);
+        self.session.update_provider(&self.config);
+
+        CommandOutcome::StartInference { pending_images }
+    }
+
+    // ═══════════════════════════════════════════════════════════
+    // Idle-mode event handling (keyboard, paste, menu, wizard, history)
+    // ═══════════════════════════════════════════════════════════
+
+    async fn handle_idle_event(&mut self, ev: Event) -> anyhow::Result<bool> {
+        match ev {
+            Event::Resize(_, _) => {
+                let _ = drain_pending_resizes(&mut self.crossterm_events);
+                scroll_past_and_reinit(
+                    &mut self.terminal,
+                    &mut self.crossterm_events,
+                    self.viewport_height,
+                )?;
+            }
+            Event::Paste(text) => {
+                self.handle_idle_paste(&text);
+            }
+            Event::Key(key) => {
+                return self.handle_idle_key(key).await;
+            }
+            _ => {}
+        }
+        Ok(true)
+    }
+
+    fn handle_idle_paste(&mut self, text: &str) {
+        let char_count = text.chars().count();
+        if matches!(self.prompt_mode, PromptMode::WizardInput { .. })
+            || char_count < input::PASTE_BLOCK_THRESHOLD
+        {
+            self.textarea.insert_str(text);
+        } else {
+            self.paste_blocks.push(input::PasteBlock {
+                content: text.to_string(),
+                char_count,
+            });
+            let label = format!("\u{1f4cb} Pasted text ({char_count} chars)");
+            emit_above(
+                &mut self.terminal,
+                Line::from(vec![
+                    Span::raw("  "),
+                    Span::styled(label, Style::default().fg(Color::Yellow)),
+                ]),
+            );
+            let preview: String = text.chars().take(80).collect();
+            let preview = preview.replace('\n', "\u{21b5}");
+            let preview = if char_count > 80 {
+                format!("{preview}\u{2026}")
+            } else {
+                preview
+            };
+            emit_above(
+                &mut self.terminal,
+                Line::from(vec![
+                    Span::raw("    "),
+                    Span::styled(preview, Style::default().fg(Color::DarkGray)),
+                ]),
+            );
+        }
+    }
+
+    async fn handle_idle_key(
+        &mut self,
+        key: crossterm::event::KeyEvent,
+    ) -> anyhow::Result<bool> {
+        if !self.menu.is_none() {
+            if let Some(consumed) = self.handle_menu_key(key).await {
+                return Ok(consumed);
+            }
+        }
+
+        match (key.code, key.modifiers) {
+            (KeyCode::Enter, m)
+                if m.contains(KeyModifiers::SHIFT) || m.contains(KeyModifiers::ALT) =>
+            {
+                self.textarea.insert_newline();
+            }
+            (KeyCode::Enter, KeyModifiers::NONE) => {
+                return self.handle_idle_enter().await;
+            }
+            (KeyCode::Up, KeyModifiers::NONE) | (KeyCode::Char('p'), KeyModifiers::CONTROL) => {
+                self.history_up();
+            }
+            (KeyCode::Down, KeyModifiers::NONE) | (KeyCode::Char('n'), KeyModifiers::CONTROL) => {
+                self.history_down();
+            }
+            (KeyCode::Esc, _) => {
+                self.textarea.select_all();
+                self.textarea.cut();
+                self.history_idx = None;
+            }
+            (KeyCode::Char('c'), m) if m.contains(KeyModifiers::CONTROL) => {
+                self.textarea.select_all();
+                self.textarea.cut();
+                self.history_idx = None;
+            }
+            (KeyCode::Char('d'), m) if m.contains(KeyModifiers::CONTROL) => {
+                if self.textarea.lines().join("").trim().is_empty() {
+                    self.should_quit = true;
+                }
+            }
+            (KeyCode::Char('l'), m) if m.contains(KeyModifiers::CONTROL) => {
+                scroll_past_and_reinit(
+                    &mut self.terminal,
+                    &mut self.crossterm_events,
+                    self.viewport_height,
+                )?;
+            }
+            (KeyCode::BackTab, _) => {
+                approval::cycle_mode(&self.shared_mode);
+            }
+            (KeyCode::Tab, KeyModifiers::NONE) => {
+                let current = self.textarea.lines().join("\n");
+                if let Some(completed) = self.completer.complete(&current) {
+                    self.textarea.select_all();
+                    self.textarea.cut();
+                    self.textarea.insert_str(&completed);
+                    self.completer.reset();
+                }
+            }
+            _ => {
+                self.history_idx = None;
+                self.completer.reset();
+                self.textarea.input(Event::Key(key));
+                self.update_reactive_menu();
+            }
+        }
+        Ok(true)
+    }
+
+    async fn handle_idle_enter(&mut self) -> anyhow::Result<bool> {
+        if matches!(self.prompt_mode, PromptMode::WizardInput { .. }) {
+            self.handle_wizard_submit().await;
+            return Ok(true);
+        }
+
+        let text = self.textarea.lines().join("\n");
+        if !text.trim().is_empty() {
+            self.textarea.select_all();
+            self.textarea.cut();
+            self.history.push(text.clone());
+            save_history(&self.history);
+            self.history_idx = None;
+            let mode = approval::read_mode(&self.shared_mode);
+            let icon = match mode {
+                ApprovalMode::Confirm => "\u{1f512}",
+                ApprovalMode::Auto => "\u{26a1}",
+            };
+            emit_above(
+                &mut self.terminal,
+                Line::from(vec![
+                    Span::styled(format!("{icon}> "), Style::default().fg(Color::Cyan)),
+                    Span::raw(text.clone()),
+                ]),
+            );
+            self.pending_command = Some(text);
+        }
+        Ok(true)
+    }
+
+    // ── Menu navigation ───────────────────────────────────────
+
+    async fn handle_menu_key(
+        &mut self,
+        key: crossterm::event::KeyEvent,
+    ) -> Option<bool> {
+        let is_up = key.code == KeyCode::Up
+            || (key.code == KeyCode::Char('k')
+                && key.modifiers.contains(KeyModifiers::CONTROL));
+        let is_down = key.code == KeyCode::Down
+            || key.code == KeyCode::Tab
+            || (key.code == KeyCode::Char('j')
+                && key.modifiers.contains(KeyModifiers::CONTROL));
+
+        if is_up {
+            self.menu_navigate(-1);
+            return Some(true);
+        }
+        if is_down {
+            self.menu_navigate(1);
+            return Some(true);
+        }
+
+        match key.code {
+            KeyCode::Enter => {
+                self.handle_menu_select().await;
+                return Some(true);
+            }
+            KeyCode::Esc => {
+                self.menu = MenuContent::None;
+                if matches!(self.prompt_mode, PromptMode::WizardInput { .. }) {
+                    self.prompt_mode = PromptMode::Chat;
+                    self.provider_wizard = None;
+                    self.textarea.select_all();
+                    self.textarea.cut();
+                }
+                return Some(true);
+            }
+            _ => {}
+        }
+        None
+    }
+
+    fn menu_navigate(&mut self, dir: i8) {
+        macro_rules! nav {
+            ($dd:expr) => {
+                if dir < 0 { $dd.up() } else { $dd.down() }
+            };
+        }
+        match &mut self.menu {
+            MenuContent::Slash(dd) => nav!(dd),
+            MenuContent::Model(dd) => nav!(dd),
+            MenuContent::Provider(dd) => nav!(dd),
+            MenuContent::Session(dd) => nav!(dd),
+            MenuContent::File { dropdown: dd, .. } => nav!(dd),
+            MenuContent::Approval { .. }
+            | MenuContent::LoopCap
+            | MenuContent::WizardTrail(_)
+            | MenuContent::None => {}
+        }
+    }
+
+    async fn handle_menu_select(&mut self) {
+        match &self.menu {
+            MenuContent::Slash(dd) => {
+                if let Some(item) = dd.selected_item() {
+                    let cmd = item.command.to_string();
+                    self.textarea.select_all();
+                    self.textarea.cut();
+                    self.textarea.insert_str(&cmd);
+                }
+            }
+            MenuContent::Model(dd) => {
+                if let Some(item) = dd.selected_item() {
+                    let model_id = item.id.clone();
+                    self.config.model = model_id.clone();
+                    self.config.model_settings.model = model_id.clone();
+                    self.config.recalculate_model_derived();
+                    {
+                        let prov = self.provider.read().await;
+                        self.config
+                            .query_and_apply_capabilities(prov.as_ref())
+                            .await;
+                    }
+                    crate::tui_wizards::save_provider(&self.config);
+                    emit_above(
+                        &mut self.terminal,
+                        Line::styled(
+                            format!("  \u{2714} Model set to: {model_id}"),
+                            Style::default().fg(Color::Green),
+                        ),
+                    );
+                    self.renderer.model = model_id;
+                }
+            }
+            MenuContent::Provider(dd) => {
+                if let Some(item) = dd.selected_item() {
+                    let key = item.key;
+                    let ptype =
+                        koda_core::config::ProviderType::from_url_or_name("", Some(key));
+                    let base_url = ptype.default_base_url().to_string();
+                    let provider_name = item.name.to_string();
+
+                    if ptype.requires_api_key() {
+                        let env_name = ptype.env_key_name().to_string();
+                        let has_key = koda_core::runtime_env::is_set(&env_name);
+                        let label = if has_key {
+                            format!("API key for {} (Enter to keep current)", ptype)
+                        } else {
+                            format!("API key for {}", ptype)
+                        };
+                        self.menu = MenuContent::WizardTrail(vec![(
+                            "Provider".into(),
+                            provider_name,
+                        )]);
+                        self.prompt_mode = PromptMode::WizardInput {
+                            label,
+                            masked: true,
+                        };
+                        self.provider_wizard = Some(ProviderWizard::NeedApiKey {
+                            provider_type: ptype,
+                            base_url,
+                            env_name,
+                        });
+                        self.textarea.select_all();
+                        self.textarea.cut();
+                    } else {
+                        self.menu = MenuContent::WizardTrail(vec![(
+                            "Provider".into(),
+                            provider_name,
+                        )]);
+                        self.prompt_mode = PromptMode::WizardInput {
+                            label: format!("{} URL", ptype),
+                            masked: false,
+                        };
+                        self.provider_wizard = Some(ProviderWizard::NeedUrl {
+                            provider_type: ptype,
+                        });
+                        self.textarea.select_all();
+                        self.textarea.cut();
+                        self.textarea.insert_str(&base_url);
+                    }
+                }
+                return;
+            }
+            MenuContent::Session(dd) => {
+                if let Some(item) = dd.selected_item() {
+                    if item.is_current {
+                        emit_above(
+                            &mut self.terminal,
+                            Line::styled(
+                                "  Already in this session.",
+                                Style::default().fg(Color::DarkGray),
+                            ),
+                        );
+                    } else {
+                        let target_id = item.id.clone();
+                        let short = item.short_id.clone();
+                        self.session.id = target_id;
+                        emit_above(
+                            &mut self.terminal,
+                            Line::from(vec![
+                                Span::styled(
+                                    "  \u{2714} ",
+                                    Style::default().fg(Color::Green),
+                                ),
+                                Span::raw("Resumed session "),
+                                Span::styled(short, Style::default().fg(Color::Cyan)),
+                            ]),
+                        );
+                    }
+                }
+            }
+            MenuContent::File { dropdown, prefix } => {
+                if let Some(item) = dropdown.selected_item() {
+                    let replacement = format!("{}@{}", prefix, item.path);
+                    self.textarea.select_all();
+                    self.textarea.cut();
+                    self.textarea.insert_str(&replacement);
+                }
+            }
+            MenuContent::Approval { .. }
+            | MenuContent::LoopCap
+            | MenuContent::WizardTrail(_)
+            | MenuContent::None => {}
+        }
+        self.menu = MenuContent::None;
+    }
+
+    // ── Wizard submit ──────────────────────────────────────────
+
+    async fn handle_wizard_submit(&mut self) {
+        let value = self.textarea.lines().join("");
+        self.textarea.select_all();
+        self.textarea.cut();
+
+        if let Some(wizard) = self.provider_wizard.take() {
+            match wizard {
+                ProviderWizard::NeedApiKey {
+                    provider_type,
+                    base_url,
+                    env_name,
+                } => {
+                    if value.is_empty() && !koda_core::runtime_env::is_set(&env_name) {
+                        emit_above(
+                            &mut self.terminal,
+                            Line::styled(
+                                "  \u{2716} No API key provided.",
+                                Style::default().fg(Color::Red),
+                            ),
+                        );
+                        self.prompt_mode = PromptMode::Chat;
+                        self.menu = MenuContent::None;
+                        return;
+                    }
+                    if !value.is_empty() {
+                        koda_core::runtime_env::set(&env_name, &value);
+                        if let Ok(mut store) = koda_core::keystore::KeyStore::load() {
+                            store.set(&env_name, &value);
+                            let _ = store.save();
+                        }
+                        let masked = koda_core::keystore::mask_key(&value);
+                        emit_above(
+                            &mut self.terminal,
+                            Line::styled(
+                                format!("  \u{2714} {env_name} set to {masked}"),
+                                Style::default().fg(Color::Green),
+                            ),
+                        );
+                    }
+                    self.apply_provider(provider_type, base_url).await;
+                }
+                ProviderWizard::NeedUrl { provider_type } => {
+                    let url = if value.is_empty() {
+                        provider_type.default_base_url().to_string()
+                    } else {
+                        value
+                    };
+                    self.apply_provider(provider_type, url).await;
+                }
+            }
+        }
+        self.prompt_mode = PromptMode::Chat;
+        self.menu = MenuContent::None;
+    }
+
+    async fn apply_provider(
+        &mut self,
+        provider_type: koda_core::config::ProviderType,
+        base_url: String,
+    ) {
+        self.config.provider_type = provider_type.clone();
+        self.config.base_url = base_url.clone();
+        self.config.model = provider_type.default_model().to_string();
+        self.config.model_settings.model = self.config.model.clone();
+        self.config.recalculate_model_derived();
+        *self.provider.write().await = koda_core::providers::create_provider(&self.config);
+        crate::tui_wizards::save_provider(&self.config);
+
+        let prov = self.provider.read().await;
+        if let Ok(models) = prov.list_models().await {
+            if let Some(first) = models.first() {
+                self.config.model = first.id.clone();
+                self.config.model_settings.model = self.config.model.clone();
+                self.config.recalculate_model_derived();
+            }
+            self.config
+                .query_and_apply_capabilities(prov.as_ref())
+                .await;
+            self.completer
+                .set_model_names(models.iter().map(|m| m.id.clone()).collect());
+        }
+        self.renderer.model = self.config.model.clone();
+        emit_above(
+            &mut self.terminal,
+            Line::styled(
+                format!(
+                    "  \u{2714} Provider: {} ({})",
+                    self.config.provider_type, self.config.model
+                ),
+                Style::default().fg(Color::Green),
+            ),
+        );
+    }
+
+    // ── History navigation ──────────────────────────────────────
+
+    fn history_up(&mut self) {
+        if !self.history.is_empty() {
+            let idx = match self.history_idx {
+                None => self.history.len() - 1,
+                Some(i) => i.saturating_sub(1),
+            };
+            self.history_idx = Some(idx);
+            self.textarea.select_all();
+            self.textarea.cut();
+            self.textarea.insert_str(&self.history[idx]);
+        }
+    }
+
+    fn history_down(&mut self) {
+        if let Some(idx) = self.history_idx {
+            if idx + 1 < self.history.len() {
+                self.history_idx = Some(idx + 1);
+                self.textarea.select_all();
+                self.textarea.cut();
+                self.textarea.insert_str(&self.history[idx + 1]);
+            } else {
+                self.history_idx = None;
+                self.textarea.select_all();
+                self.textarea.cut();
+            }
+        }
+    }
+
+    // ── Reactive menu updates ───────────────────────────────────
+
+    fn update_reactive_menu(&mut self) {
+        let after_input = self.textarea.lines().join("\n");
+        let trimmed = after_input.trim_end();
+
+        if trimmed.starts_with('/') && !trimmed.contains(' ') {
+            if let Some(dd) = crate::widgets::slash_menu::from_input(
+                crate::completer::SLASH_COMMANDS,
+                trimmed,
+            ) {
+                self.menu = MenuContent::Slash(dd);
+            } else if matches!(self.menu, MenuContent::Slash(_)) {
+                self.menu = MenuContent::None;
+            }
+        } else if let Some(at_pos) = crate::completer::find_last_at_token(trimmed) {
+            let partial = &trimmed[at_pos + 1..];
+            let prefix = &trimmed[..at_pos];
+            let matches =
+                crate::completer::list_path_matches_public(&self.project_root, partial);
+            if !matches.is_empty() {
+                let items: Vec<crate::widgets::file_menu::FileItem> = matches
+                    .iter()
+                    .map(|p| crate::widgets::file_menu::FileItem {
+                        path: p.clone(),
+                        is_dir: p.ends_with('/'),
+                    })
+                    .collect();
+                let dd = crate::widgets::dropdown::DropdownState::new(
+                    items,
+                    "\u{1f4c2} Files",
+                );
+                self.menu = MenuContent::File {
+                    dropdown: dd,
+                    prefix: prefix.to_string(),
+                };
+            } else if matches!(self.menu, MenuContent::File { .. }) {
+                self.menu = MenuContent::None;
+            }
+        } else if matches!(
+            self.menu,
+            MenuContent::Slash(_) | MenuContent::File { .. }
+        ) {
+            self.menu = MenuContent::None;
+        }
     }
 }
